+++ src/libutil/error.cc
@@ -2,6 +2,7 @@
 
 #include "nix/util/error.hh"
 #include "nix/util/environment-variables.hh"
+#include "nix/util/exit.hh"
 #include "nix/util/signals.hh"
 #include "nix/util/terminal.hh"
 #include "nix/util/position.hh"
@@ -453,6 +454,35 @@ void unreachable(std::source_location loc)
     if (n < 0)
         panic("Unexpected condition and could not format error message");
     panic(std::string_view(buf, std::min(static_cast<int>(sizeof(buf)), n)));
+}
+
+int handleExceptions(const std::string & programName, std::function<void()> fun)
+{
+    ReceiveInterrupts receiveInterrupts; // FIXME: need better place for this
+
+    ErrorInfo::programName = baseNameOf(programName);
+
+    std::string error = ANSI_RED "error:" ANSI_NORMAL " ";
+    try {
+        fun();
+    } catch (Exit & e) {
+        return e.status;
+    } catch (UsageError & e) {
+        logError(e.info());
+        printError("Try '%1% --help' for more information.", programName);
+        return 1;
+    } catch (BaseError & e) {
+        logError(e.info());
+        return e.info().status;
+    } catch (std::bad_alloc & e) {
+        printError(error + "out of memory");
+        return 1;
+    } catch (std::exception & e) {
+        printError(error + e.what());
+        return 1;
+    }
+
+    return 0;
 }
 
 } // namespace nix
